Refactor hpegl-provider-lib/pkg/token packages to eliminate repeated code
